backend/requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# PDF Processing
pypdf>=5.0.0

# Vector Store
faiss-cpu>=1.7.0

# BM25 lexical search
rank-bm25>=0.2.2

# Google AI
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0

# API
flask>=3.0.0
flask-cors>=4.0.0

# Server WSGI pour production
gunicorn>=21.0.0
